add `pkg/id` package
`id.New()` produces a 14 character NanoID identifier from the following
set of characters:

"0123456789abcdefghijkmnopqrstuvwxyz"

Note that lowercase "L" is not in the list. This is because that
character is hard to distinguish from the number 1.

According to NanoID, this generator scheme can produce 100K IDs per hour
for 10 years before there is a 1% chance of collision.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>